Factor our redundant code
