+++ src/tools/concurrency/nodes/TraceActorContext.java
@@ -1,34 +1,14 @@
 package tools.concurrency.nodes;
 
 import com.oracle.truffle.api.dsl.Specialization;
-import com.oracle.truffle.api.nodes.Node;
 
-import som.vm.VmSettings;
 import tools.concurrency.ActorExecutionTrace;
 import tools.concurrency.ActorExecutionTrace.ActorTraceBuffer;
 import tools.concurrency.ByteBuffer;
-import tools.concurrency.TracingActivityThread;
 import tools.concurrency.TracingActors.TracingActor;
 
 
-public abstract class TraceActorContext extends Node {
-  public abstract void execute(TracingActor actor);
-
-  protected static boolean smallIds() {
-    return VmSettings.TRACE_SMALL_IDS;
-  }
-
-  protected static boolean byteId(final TracingActor actor) {
-    return (actor.getActorId() & 0xFFFFFF00) == 0;
-  }
-
-  protected static boolean shortId(final TracingActor actor) {
-    return (actor.getActorId() & 0xFFFF0000) == 0;
-  }
-
-  protected static boolean threeByteId(final TracingActor actor) {
-    return (actor.getActorId() & 0xFF000000) == 0;
-  }
+public abstract class TraceActorContext extends TraceNode {
 
   @Specialization(guards = {"smallIds()", "byteId(actor)"})
   public static void traceByteId(final TracingActor actor) {
@@ -64,16 +44,5 @@ public abstract class TraceActorContext extends Node {
   private static ByteBuffer getStorage() {
     ActorTraceBuffer buffer = getCurrentBuffer();
     return buffer.getStorage();
-  }
-
-  private static TracingActivityThread getThread() {
-    Thread current = Thread.currentThread();
-    assert current instanceof TracingActivityThread;
-    return (TracingActivityThread) current;
-  }
-
-  private static ActorTraceBuffer getCurrentBuffer() {
-    TracingActivityThread t = getThread();
-    return (ActorTraceBuffer) t.getBuffer();
   }
 }
